chore: move http_client from /core to /service_layer
